ai-zoo-discord-bots/requirements.txt
# Discord API
discord.py>=2.0.0

# HTTP client for API requests
aiohttp>=3.8.0

# Environment variable management
python-dotenv>=0.19.0

# Development tools (optional)
watchgod>=0.7.0  # For auto-reloading in development

# Typing extensions
typing-extensions>=4.0.0

# Date and time handling
pytz>=2021.3

# Logging enhancements
colorlog>=6.6.0

# For Notion API
notion-client>=1.0.0

# For OpenAI API
openai>=0.27.0

# For Anthropic API
anthropic>=0.2.0
